requirements.txt
Django==3.0.6
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.8.6

